feat: add GameProgress model and update service methods to use it
